map \da to add the date
This adds a pound-sign, then the date on a new line, then moves up one
line and joins the line below. Should work great for journal-style
markdown where the H1 tag is the timestamp, and any config file whose
syntax allows comments that begin the line with a pound sign.