Use TeX Gyre fonts.
Apparently in Debian, the fonts labeled "Nimbus Roman No9 L" and "Nimbus
Sans L" are from the gsfonts package, in the Type1 format. Google Chrome
rendered no text, but left the right amount of space, wherever these
fonts were called for. The TeX Gyre fonts are basically the same thing,
but in a more compatible format. TeX Gyre is in the package
fonts-texgyre.